unir-backend/requirements.txt
importlib
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
PyJWT
mysql-connector-python
configparser